Initial commit
I've hosted the project just on S3  (since 2013) because this is
an small project. I'm moving it to Github now.

Motivation:
----------

I developed this small project to demostrate my skillsets with
javascript development and electrical engineering too for a job
interview I had in 2013 (Which helped me to land an awesome job
by the way).